Protected admin API
Enhances membership deletion logic with additional checks
Improves UI consistency in table rows and tooltips

+++ src/features/membership/actions/delete-membership.ts
@@ -12,15 +12,13 @@ export const deleteMembership = async ({
   organizationId: string;
   userId: string;
 }) => {
-  const auth = await getAuthOrRedirect();
+  const { user } = await getAuthOrRedirect();
 
   const membershipsResult = await getMemberships(organizationId);
   const memberships = membershipsResult.organizationMemberships;
 
+  // Check if the user is trying to delete the last membership
   const isLastMembership = (memberships ?? []).length <= 1;
-
-  const contextAwareToast =
-    userId === auth.user.id ? "Leaving Organization..." : "Deleting member...";
 
   if (isLastMembership) {
     return toActionState(
@@ -29,6 +27,46 @@ export const deleteMembership = async ({
     );
   }
 
+  // Check if the membership exists
+  const targetMembership = (memberships ?? []).find(
+    (membership) => membership.userId === userId,
+  );
+
+  if (!targetMembership) {
+    return toActionState("ERROR", "Membership not found");
+  }
+
+  // Check if the user is deleting the last admin
+  const adminMemberships = (memberships ?? []).filter(
+    (membership) => membership.membershipRole === "ADMIN",
+  );
+
+  const removesAdmin = targetMembership.membershipRole === "ADMIN";
+  const isLastAdmin = adminMemberships.length <= 1;
+
+  if (removesAdmin && isLastAdmin) {
+    return toActionState(
+      "ERROR",
+      "You cannot delete the last admin of an organization.",
+    );
+  }
+
+  // Check if the user is authorized to delete the membership
+  const myMembership = (memberships ?? []).find(
+    (membership) => membership.userId === user.id,
+  );
+
+  const isMyself = userId === user.id;
+  const isAdmin = myMembership?.membershipRole === "ADMIN";
+
+  if (!isAdmin && !isMyself) {
+    return toActionState(
+      "ERROR",
+      "You can only delete memberships as an admin",
+    );
+  }
+
+  // Everything is checked. Proceed to delete the membership
   await prisma.membership.delete({
     where: {
       membershipId: {
@@ -37,6 +75,10 @@ export const deleteMembership = async ({
       },
     },
   });
+
+  const contextAwareToast = isMyself
+    ? "You successfully left the organization."
+    : "Membership deleted";
 
   return toActionState("SUCCESS", `${contextAwareToast}`);
 };
